refactoring of buildings rendering much like obstacles rendering (moving to a controller, renderer and grid builder + text texture handler & territories handler
