+++ wos-interactive-map-front/src/components/MapContainer.jsx
@@ -1,5 +1,5 @@
 import React, { use, useEffect, useRef } from 'react';
-import { initScene, renderObstacles, syncBuildings } from '../three/main.js'; // adjust path as needed
+import { initScene, renderObstacles, renderBuildings } from '../three/main.js'; // adjust path as needed
 
 
 export default function MapContainer({buildings, onCellSelect,setLoading }) {
@@ -17,7 +17,7 @@ export default function MapContainer({buildings, onCellSelect,setLoading }) {
   useEffect(()=>{
     if (buildings) {
       if (!buildings || buildings.length === 0) return;
-      syncBuildings(buildings, setLoading);
+      renderBuildings(buildings, setLoading);
     }
   },[buildings])
